+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Journal des modifications
 
 9.2.1
 -----
+- [FIX] Module Admission : Notifications non reçues 
 - [FIX] Dysfonctionnements autour des comptes utilisateurs (accès à la page perturbé suite au passage à unicaen/utilisateur, bug dans la création de compte+individu et l'envoi de la notif).
 - [FIX] Envoi de la notif 'Validation du dépôt de la thèse corrigée' à un président du jury via la page Admin > Présidents.
 - [FIX] Utilisation de UrlService (plutôt que UrlHelper) pour générer l'URL d'usurpation d'identité à partir d'un individu.